app/requirements.txt
cassandra-driver==3.19.0
Django==2.2.6
django-cassandra-engine==1.5.5
djangorestframework==3.10.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
